requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0fastapi==0.119.0
uvicorn==0.37.0
SQLAlchemy==2.0.44
databases==0.9.0
asyncpg==0.30.0
python-jose==3.5.0
python-dotenv==1.1.1
celery==5.5.3
redis==5.2.1
websockets==15.0.1
httpx==0.28.1
fastapi==0.119.0
uvicorn==0.37.0
SQLAlchemy==2.0.44
databases==0.9.0
asyncpg==0.30.0
python-jose==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.1
celery==5.5.3
redis==5.2.1
websockets==15.0.1
httpx==0.28.1
black==25.9.0
mypy==1.7.1
